Extend clang-tidy checks into  tools, utils, and python_bindings
(and also fix the reported findings)

Note that this also extends the testing to src/autoschedulers, but *not* to src/runtime (yet). It would likely be desirable to do so at some point, but requires a bit more massaging of the build scripts.

Extending to apps/ and test/ might be nice, but would require running all Generators to produce the needed .h files.